PROBE/CompGCN/requirements.txt
pytorch >= 2.6.0
numpy >= 2.2.4
pykeen >= 1.11.0 